app/requirements.txt
streamlit==1.8.1
Pillow==9.2.0
matplotlib==3.5.1
seaborn==0.11.2
pandas==1.4.1
numpy==1.21.2
folium==0.12.1.post1
streamlit-folium==0.6.13

